moved UserProfileViewModelFactory back to original directory to fix conflict
